Changed BigInteger to Long, to improve the server performance.
git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@517448 13f79535-47bb-0310-9956-ffa450edef68

+++ bootstrap-plugin/src/main/java/org/apache/directory/server/core/bootstrap/plugin/BootstrapPlugin.java
@@ -24,7 +24,6 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -785,7 +784,7 @@ public class BootstrapPlugin extends AbstractMojo
     
     private final boolean hasEntry( LdapDN dn ) throws NamingException
     {
-        BigInteger id = store.getEntryId( dn.toNormName() );
+        Long id = store.getEntryId( dn.toNormName() );
         if ( id == null )
         {
             return false;
