+++ core/src/main/java/org/apache/directory/server/core/partition/impl/btree/BTreeSearchResult.java
@@ -20,8 +20,6 @@
 package org.apache.directory.server.core.partition.impl.btree;
 
 
-import java.math.BigInteger;
-
 import javax.naming.directory.Attributes;
 import javax.naming.directory.SearchResult;
 
@@ -39,7 +37,7 @@ public class BTreeSearchResult extends SearchResult
     private static final long serialVersionUID = 3976739172700860977L;
 
     /** the primary key used for the resultant entry */
-    private final BigInteger id;
+    private final Long id;
 
 
     // ------------------------------------------------------------------------
@@ -54,7 +52,7 @@ public class BTreeSearchResult extends SearchResult
      * @param obj the object if any
      * @param attrs the attributes of the entry
      */
-    public BTreeSearchResult(BigInteger id, String name, Object obj, Attributes attrs)
+    public BTreeSearchResult(Long id, String name, Object obj, Attributes attrs)
     {
         super( name, obj, attrs );
         this.id = id;
@@ -70,7 +68,7 @@ public class BTreeSearchResult extends SearchResult
      * @param attrs the attributes of the entry
      * @param isRelative whether or not the name is relative to the base
      */
-    public BTreeSearchResult(BigInteger id, String name, Object obj, Attributes attrs, boolean isRelative)
+    public BTreeSearchResult(Long id, String name, Object obj, Attributes attrs, boolean isRelative)
     {
         super( name, obj, attrs, isRelative );
         this.id = id;
@@ -86,7 +84,7 @@ public class BTreeSearchResult extends SearchResult
      * @param obj the object if any
      * @param attrs the attributes of the entry
      */
-    public BTreeSearchResult(BigInteger id, String name, String className, Object obj, Attributes attrs)
+    public BTreeSearchResult(Long id, String name, String className, Object obj, Attributes attrs)
     {
         super( name, className, obj, attrs );
         this.id = id;
@@ -103,7 +101,7 @@ public class BTreeSearchResult extends SearchResult
      * @param attrs the attributes of the entry
      * @param isRelative whether or not the name is relative to the base
      */
-    public BTreeSearchResult(BigInteger id, String name, String className, Object obj, Attributes attrs,
+    public BTreeSearchResult(Long id, String name, String className, Object obj, Attributes attrs,
         boolean isRelative)
     {
         super( name, className, obj, attrs, isRelative );
@@ -116,7 +114,7 @@ public class BTreeSearchResult extends SearchResult
      * 
      * @return Returns the id.
      */
-    public BigInteger getId()
+    public Long getId()
     {
         return id;
     }
