+++ core/src/main/java/org/apache/directory/server/core/partition/impl/btree/BTreeSearchResultEnumeration.java
@@ -20,7 +20,6 @@
 package org.apache.directory.server.core.partition.impl.btree;
 
 
-import java.math.BigInteger;
 import java.util.Iterator;
 
 import javax.naming.NamingEnumeration;
@@ -105,11 +104,11 @@ public class BTreeSearchResultEnumeration implements SearchResultEnumeration
     {
         IndexRecord rec = ( IndexRecord ) underlying.next();
         Attributes entry;
-        String name = partition.getEntryUpdn( (BigInteger)rec.getEntryId() );
+        String name = partition.getEntryUpdn( (Long)rec.getEntryId() );
 
         if ( null == rec.getAttributes() )
         {
-            rec.setAttributes( partition.lookup( (BigInteger)rec.getEntryId() ) );
+            rec.setAttributes( partition.lookup( (Long)rec.getEntryId() ) );
         }
 
         if ( attrIds == null )
@@ -235,7 +234,7 @@ public class BTreeSearchResultEnumeration implements SearchResultEnumeration
             }
         }
 
-        BTreeSearchResult result = new BTreeSearchResult( (BigInteger)rec.getEntryId(), name, null, entry );
+        BTreeSearchResult result = new BTreeSearchResult( (Long)rec.getEntryId(), name, null, entry );
         result.setRelative( false );
         return result;
     }
